Break out more test classes to their own files
